refactor: constants와 utils를 lib폴더에 분리
